Remove .NET 3.1 example
