--- examples/WebApp.NextVersion/appsettings.json
@@ -1,18 +1,0 @@
-{
-  "Logging": {
-    "LogLevel": {
-      "Default": "Information",
-      "Microsoft.AspNetCore": "Warning"
-    },
-    "Telegram": {
-      "LogLevel": {
-        "Default": "Error"
-      },
-      "AccessToken": "1234567890:AAAaaAAaa_AaAAaa-AAaAAAaAAaAaAaAAAA",
-      "ChatId": "-0000000000000",
-      "UseEmoji": "true",
-      "Source": "WebApp.NextVersion"
-    }
-  },
-  "AllowedHosts": "*"
-}
