+++ examples/WebApp/appsettings.json
@@ -2,18 +2,16 @@
   "Logging": {
     "LogLevel": {
       "Default": "Information",
-      "Microsoft": "Warning",
-      "Microsoft.Hosting.Lifetime": "Information"
+      "Microsoft.AspNetCore": "Warning"
     },
     "Telegram": {
       "LogLevel": {
-        "Default": "Error",
-        "WebApp.Controllers": "Warning"
+        "Default": "Error"
       },
       "AccessToken": "1234567890:AAAaaAAaa_AaAAaa-AAaAAAaAAaAaAaAAAA",
       "ChatId": "-0000000000000",
       "UseEmoji": "true",
-      "Source": "WebApp"
+      "Source": "WebApp.NextVersion"
     }
   },
   "AllowedHosts": "*"
